Pulling own events out into separate component
