+++ src/pages/ProfilePage.jsx
@@ -1,23 +1,17 @@
 import useSelf from "../hooks/use-self";
 import useAuth from "../hooks/use-auth";
-import useMyEvents from "../hooks/use-myevents";
-import useEvents from "../hooks/use-events";
-import { convertLocalDateTime } from "../utlities/convertLocalDateTime";
 import NotFound404Page from "../components/NotFound404Page/NotFound404Page";
 import Button from "../components/Buttton/Button";
 import Profile from "../components/Profile/Profile.jsx";
+import MyEventsComponent from "../components/MyEvents/MyEvents";
 import UserUpdateForm from "../components/UpdateProfileForm/UpdateProfileForm";
 import { useState } from "react";
-import { useNavigate } from "react-router-dom";
 
 function ProfilePage() {
   
   const { auth, setAuth } = useAuth();
   const { self, isLoading, error } = useSelf();
   const [ editing, setEditing ] = useState(false);
-  const [myEvents, myEventsLoading, myEventsError] = useMyEvents();
-  const { events, isEventsLoading, isEventsError } = useEvents();
-  const navigate = useNavigate();
   
   if(!auth.token) {
     return <NotFound404Page/>
@@ -30,27 +24,11 @@ function ProfilePage() {
   if (error) {
     return <p>{error.message}</p>;
   }
-  
-  const handleEventClick = (event) => {
-    let eventid = event.target.value;
-    navigate(`/events/${eventid}`);
-  };
 
   const handleUpdate = (event) => {
     event.preventDefault();
     setEditing(!editing);
     }
-
-  const myEventIds = [];
-  for (let myEvent in myEvents) {
-    myEventIds.push(myEvents[myEvent]["event_id"]);
-  }
-  const registeredEvents = [];
-  myEventIds.forEach(function (eventid) {
-    registeredEvents.push(events.filter((event) => event["id"] == eventid));
-  });
-  const myRegisteredEvents = registeredEvents.flatMap((event) => event);
-
 
   return (
     <div>
@@ -58,26 +36,8 @@ function ProfilePage() {
       {
         editing == false ? <><Profile /> 
         <Button text={"Update Details"} btnClass = "btn-info" onClick={handleUpdate}/>
-        <div>
-          {myRegisteredEvents.map((event, key) => {
-            return (
-              <div className="event-single-card" key={key}>
-                <div>
-                  <p>Date: {convertLocalDateTime(event.start_date)}</p>
-                  <p>Title: {event.title}</p>
-                  <p>Location: {event.location}</p>
-                </div>
-                <button
-                  className="button"
-                  onClick={handleEventClick}
-                  value={event.id}>
-                  Find out more
-                </button>
-              </div>
-            );
-          })}
-        </div>
-        </>
+        <br></br>
+        <MyEventsComponent /></>
         : <>
         <UserUpdateForm editing={editing} setEditing={setEditing}/>
           <Button text={"Return to profile view"} btnClass = "btn-info update-button" onClick={handleUpdate}/>
